requirements.txt
RPi.GPIO
RPi.bme280
smbus2
